Split tests into parsing and encoding tests
